+++ app/src/main/java/de/baumann/quitsmoking/helper/Activity_EditNote.java
@@ -1,11 +1,16 @@
 package de.baumann.quitsmoking.helper;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.database.Cursor;
+import android.graphics.Bitmap;
+import android.net.Uri;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
+import android.provider.MediaStore;
 import android.support.annotation.NonNull;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
@@ -16,7 +21,6 @@ import android.view.MenuItem;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.inputmethod.InputMethodManager;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
@@ -24,6 +28,9 @@ import android.widget.ImageButton;
 import android.widget.ListAdapter;
 import android.widget.TextView;
 
+import com.mvc.imagepicker.ImagePicker;
+
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -37,7 +44,6 @@ public class Activity_EditNote extends AppCompatActivity {
     private Button attachment;
     private ImageButton attachmentRem;
     private ImageButton attachmentCam;
-    private ImageButton attachmentPic;
     private EditText titleInput;
     private EditText textInput;
     private SharedPreferences sharedPref;
@@ -68,7 +74,6 @@ public class Activity_EditNote extends AppCompatActivity {
         attachmentRem = (ImageButton) findViewById(R.id.button_rem);
         attachment = (Button) findViewById(R.id.button_att);
         attachmentCam = (ImageButton) findViewById(R.id.button_cam);
-        attachmentPic = (ImageButton) findViewById(R.id.button_pic);
 
         String att = getString(R.string.note_attachment) + ": " + attName;
 
@@ -76,19 +81,16 @@ public class Activity_EditNote extends AppCompatActivity {
             attachment.setText(R.string.choose_att);
             attachmentRem.setVisibility(View.GONE);
             attachmentCam.setVisibility(View.VISIBLE);
-            attachmentPic.setVisibility(View.VISIBLE);
         } else {
             attachment.setText(att);
             attachmentRem.setVisibility(View.VISIBLE);
             attachmentCam.setVisibility(View.GONE);
-            attachmentPic.setVisibility(View.GONE);
         }
         File file2 = new File(file);
         if (!file2.exists()) {
             attachment.setText(R.string.choose_att);
             attachmentRem.setVisibility(View.GONE);
             attachmentCam.setVisibility(View.VISIBLE);
-            attachmentPic.setVisibility(View.VISIBLE);
         }
 
         titleInput = (EditText) findViewById(R.id.note_title_input);
@@ -133,18 +135,6 @@ public class Activity_EditNote extends AppCompatActivity {
                 attachment.setText(R.string.choose_att);
                 attachmentRem.setVisibility(View.GONE);
                 attachmentCam.setVisibility(View.VISIBLE);
-                attachmentPic.setVisibility(View.VISIBLE);
-            }
-        });
-
-        attachmentPic.setOnClickListener(new View.OnClickListener() {
-
-            @Override
-            public void onClick(View arg0) {
-                Intent intent = new Intent(Activity_EditNote.this, Activity_images.class);
-                intent.setAction("intent_note");
-                intent.addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION);
-                startActivity(intent);
             }
         });
 
@@ -152,16 +142,7 @@ public class Activity_EditNote extends AppCompatActivity {
 
             @Override
             public void onClick(View arg0) {
-
-                File f = helper_main.newFile();
-                final String fileName = f.getAbsolutePath();
-                sharedPref.edit().putString("handleTextAttachment", fileName).apply();
-
-                InputMethodManager imm = (InputMethodManager)getSystemService(Context.INPUT_METHOD_SERVICE);
-                imm.hideSoftInputFromWindow(titleInput.getWindowToken(), 0);
-                Intent intent = new Intent(Activity_EditNote.this, Popup_camera.class);
-                intent.addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION);
-                startActivity(intent);
+                onPickImage();
             }
         });
 
@@ -380,6 +361,39 @@ public class Activity_EditNote extends AppCompatActivity {
     }
 
     @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+
+        Bitmap bitmap = ImagePicker.getImageFromResult(this, requestCode, resultCode, data);
+
+        if (bitmap != null) {
+            // CALL THIS METHOD TO GET THE URI FROM THE BITMAP
+            Uri tempUri = getImageUri(getApplicationContext(), bitmap);
+            String path = getRealPathFromURI(tempUri);
+            sharedPref.edit().putString("handleTextAttachment", path).apply();
+        }
+    }
+
+    private Uri getImageUri(Context inContext, Bitmap inImage) {
+        ByteArrayOutputStream bytes = new ByteArrayOutputStream();
+        inImage.compress(Bitmap.CompressFormat.JPEG, 100, bytes);
+        String path = MediaStore.Images.Media.insertImage(inContext.getContentResolver(), inImage, "Title", null);
+        return Uri.parse(path);
+    }
+
+    private String getRealPathFromURI(Uri uri) {
+        @SuppressLint("Recycle") Cursor cursor = getContentResolver().query(uri, null, null, null, null);
+        assert cursor != null;
+        cursor.moveToFirst();
+        int idx = cursor.getColumnIndex(MediaStore.Images.ImageColumns.DATA);
+        return cursor.getString(idx);
+    }
+
+    private void onPickImage() {
+        // Click on image button
+        ImagePicker.pickImage(this, "Select your image:");
+    }
+
+    @Override
     protected void onResume() {
         super.onResume();    //To change body of overridden methods use File | Settings | File Templates.
 
@@ -396,19 +410,16 @@ public class Activity_EditNote extends AppCompatActivity {
             attachment.setText(R.string.choose_att);
             attachmentRem.setVisibility(View.GONE);
             attachmentCam.setVisibility(View.VISIBLE);
-            attachmentPic.setVisibility(View.VISIBLE);
         } else {
             attachment.setText(att);
             attachmentRem.setVisibility(View.VISIBLE);
             attachmentCam.setVisibility(View.GONE);
-            attachmentPic.setVisibility(View.GONE);
         }
         File file2 = new File(file);
         if (!file2.exists()) {
             attachment.setText(R.string.choose_att);
             attachmentRem.setVisibility(View.GONE);
             attachmentCam.setVisibility(View.VISIBLE);
-            attachmentPic.setVisibility(View.VISIBLE);
         }
     }